Started refactoring the project around
