--- C:\Jython\testlog4j.log
@@ -1,1 +1,0 @@
-DEBUG main 2019-09-28 21:40:07.046000 - We be debuggin
