+++ src/main/java/DbDiff.java
@@ -33,10 +33,10 @@ public class DbDiff {
         formatter.printHelp("CLITester", options);
 
         /** Should be able to use this example straing from Yaml file for dbtype */
-        Enum dbtype = Enum.valueOf(DbConn.DbType.class, "SYBASE");
+        //Enum dbtype = Enum.valueOf(DbConn.DbType.class, "SYBASE");
         // This should be all you need to get a jdbc connection
         try {
-            DbConn sybaseConn = new DbConn(DbConn.DbType.SYBASE, "sa", sybasePassword, "dbsybase", "5000", "master");
+          //  DbConn sybaseConn = new DbConn(DbConn.DbType.SYBASE, "sa", sybasePassword, "dbsybase", "5000", "master");
             //DbConn oracleConn = new DbConn(DbConn.DbType.ORACLE, "system", "Docker12345", "dboracle", "1542", "dev");
         } catch (Exception e) {
             System.out.println(e);
