+++ src/main/java/Utils/JLogger.java
@@ -1,3 +1,5 @@
+package Utils;
+
 import org.apache.log4j.*;
 
 // https://jython.readthedocs.io/en/latest/appendixB/#logging
